Lissajous polynomials

The Lissajous polynomials are the implicit (variety) descriptions of the image of the parametric map x = cos(m t + p), y = sin(n t + q).

lissajous(m, n, p, q, digits = 3)

Arguments

m, n, p, q
Trigonometric coefficients, see examples for description

digits
The number of digits to round coefficients to, see round.mpoly(). This is useful for cleaning terms that are numerically nonzero, but should be.

Value

a mpoly object

See also

Examples

lissajous(3, 2,  -pi/2, 0)

#> -4 x^2  +  4 x^4  +  9 y^2  -  24 y^4  +  16 y^6

lissajous(4, 3,  -pi/2, 0)

#> 9 x^2  -  24 x^4  +  16 x^6  -  16 y^2  +  80 y^4  -  128 y^6  +  64 y^8